+++ libHDiffPatch/HPatch/hpatch_mt/_hcache_old_mt.h
@@ -1,4 +1,4 @@
-//  houtput_mt.h
+//  _hcache_old_mt.h
 //  hpatch
 /*
  The MIT License (MIT)
@@ -25,29 +25,25 @@
  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
  OTHER DEALINGS IN THE SOFTWARE.
 */
-#ifndef houtput_mt_h
-#define houtput_mt_h
-#include "hpatch_mt.h"
+#ifndef _hcache_old_mt_h
+#define _hcache_old_mt_h
+#include "_hpatch_mt.h"
 #ifdef __cplusplus
 extern "C" {
 #endif
 #if (_IS_USED_MULTITHREAD)
 
-struct houtput_mt_t;
+size_t               hcache_old_mt_t_size();
 
-size_t               houtput_mt_t_memSize();
-
-// create a new houtput_mt_t* wrapper base_stream;
-//   start a thread to write data to base_stream
-struct houtput_mt_t* houtput_mt_open(struct houtput_mt_t* pmem,size_t memSize,struct hpatch_mt_t* h_mt,
-                                     const hpatch_TStreamOutput* base_stream,hpatch_StreamPos_t curWritePos);
-// write data to base_stream
-hpatch_BOOL          houtput_mt_write(struct houtput_mt_t* self,struct hpatch_TWorkBuf* data);
-// wait write data thread end and free self
-hpatch_BOOL          houtput_mt_close(struct houtput_mt_t* self);
+// create a new hpatch_TStreamInput* wrapper old_stream;
+//   start a thread to read data from old_stream
+hpatch_TStreamInput* hcache_old_mt_open(void* pmem,size_t memSize,struct hpatch_mt_t* h_mt,struct hpatch_TWorkBuf* freeBufList,
+                                        const hpatch_TStreamInput* old_stream,sspatch_coversListener_t** out_coversListener,
+                                        sspatch_coversListener_t* nextCoverlistener);
+hpatch_BOOL          hcache_old_mt_close(const hpatch_TStreamInput* hcache_old_mt_stream);
 
 #endif
 #ifdef __cplusplus
 }
 #endif
-#endif //houtput_mt_h
+#endif //_hcache_old_mt_h
