+++ libHDiffPatch/HPatch/hpatch_mt/_hinput_mt.c
@@ -1,4 +1,4 @@
-//  hinput_mt_t.c
+//  _hinput_mt.c
 //  hpatch
 /*
  The MIT License (MIT)
@@ -25,8 +25,8 @@
  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
  OTHER DEALINGS IN THE SOFTWARE.
 */
-#include "hinput_mt.h"
-#include "patch_private_mt.h"
+#include "_hinput_mt.h"
+#include "_patch_private_mt.h"
 #if (_IS_USED_MULTITHREAD)
 
 typedef struct hinput_mt_t{
